Relational arg constraints: For bools, consider 'false' missing
Right now, all bools are implicitly default false. This means they are
never considered "missing" for relational constraint purposes. So for
example:

  args:
    authenticate bool
    token string

    authenticate mutually requires token

  if authenticate:
    print("Token:", token)

In this example, prior to this commit, users can define *just the
token*, and leave 'authenticate' false. Because bool args are default
false and don't count as missing, the 'requirement' validation will
always consider a bool arg 'present' or 'defined', and so the
requirement really has not impact.

Functionally, from a user-perspective, if you write something like the
above, what you're really after is for 'authenticate' to be true
whenever 'token' is given. This means we must handle the validation
logic for bools a little differently which this commit does. For
validation, if a bool arg is false, then it counts as 'not supplied' for
the purposes of excludes/requires.

Addresses one of the two major follow-ups required in #14.