Add rad benchmark folder
It'd be good to begin benchmarking rad. This commit adds some basic
infra for such things. We use hyperfine to conduct the actual
benchmarking, adding some RSL scripts on top to facilitate running the
benchmarks and reporting out the results.

The outputting right now is fairly primitive -- we'll just append a
basic table into the benchmark/README.md which will contain a history of
results over time. We also maintain the raw hyperfine JSON reports
though, so we can reprocess in the future and represent differently, if
so desired (including comparing benchmarks over time).