Add new benchmark after recent expr overhaul
As predicted, we've taken a performance hit. Actually larger than I
expected. I still suspect the biggest drawdown in performance atm is the
fact that we are doing a string comparison every single time we visit a
node. If we could instead do an int/id comparison, I *think* this would
give us a big speedup.