Implement new switch statement grammar
Previously, switch statements were very limited. You could only have a
single line of exprs per case. There was also no default branch.

With this commit, we update the the interpreter to handle the new
grammar which allows full blocks and a default branch. We borrow the
'yield' keyword from Java for returning values in the case of switch
expressions. Here's an example of a mixed switch expression:

  a = switch num:
    case 1, 2 -> 10
    case 3:
      print(30, 40)
      yield 30
    default -> 50

'->' allows directly returning a value in a single line. Colon newline
is used for blocks. When wanting to return values from blocks, you use
'yield'.

I thought about some alternatives here. For Rust, they did this sorta
pattern instead:

  case 3:
    print(30, 40)
    (30)

i.e. the value of the last statement/expr of a block is what gets
returned. I found this neat, but it makes more sense in Rust where this
sort of idea is leveraged in several places. For RSL, this would be the
only place (at the moment), and so it'd just appear like a surprising
syntax imo. Yield is a little more familiar.