Allow continuing/breaking from switch expression
Previous to this commit, this was not allowed:

  for i in range(5):
    a = switch i:
      case 0:
	    yield 10
      case 1:
	    break

because we'd require 'case 1' to yield as it is an expression, not
realizing that we'll never need the yield because one of the statements
in the block "cancels" the assigment.

We fix that with this commit -- you can now break or continue.

While here, I noticed another bug -- I was not remembering to reset the
'continuing' flag when iterating through lists/strings. So once we
continue'd once, we'd never run the for loop again. We fix that while
here and add a test for coverage.