Add function 'write_file'
So far, the approach for writing to files has been to doing something
like

  $!`echo -n '{content}' > {path}`

This *works* but is clearly a workaround of the fact that RSL doesn't
offer an in-built utility for this, so let's patch that up. It also will
probably run into issues if the contents contain single quotes.

We keep the function simple:

  write_file(path, content, append = false)

We might want some fancier options like writing bytes.