Add functions 'bold', 'italic', 'underline'
Involves a little bit of rejigging -- so far, all text attributes have
been mutually exclusive colors. Bold/italic/underline is also modeled as
'Attributes' in amterp/color, but the difference is that they are not
mutually exclusive from one another and colors, so Rad now needs to be
able to have several attributes on segments, which we implement.

After this change, you can now have a red underlined piece of text, etc.

We don't implement these attributes for rad tables, however, as that
requires updating amterp/go-tbl a bit, and tbh, I'm rethinking doing any
of this logic in that library, and perhaps instead doing it in rad
instead using the field modifier 'map' logic.