Function refactor: Introduce PositionalArgSchema
The current approach assumes that the args to a given function is
enumerable. Some functions e.g. print() can take an infinite number of
args, so we can't properly encode that atm.

This refactor doesn't change behavior, we're still using the enumerable
approach, but we're set up to introduce alternative validators.