Allow passing lists, maps to 'display' blocks
Prior to this commit, in order to leverage the "rad block json
extraction algo" i.e. the algo which extracts specified json paths from
a json blob, you'd need to either be getting the json from an HTTP
request done through the rad block, or utilize the --mock-response
(intended for testing) global flag.

There are cases where you may with to source the data differently.

For example, you may wish to read the JSON data from a file (without
using --mock-response). Or, you may wish to handle the HTTP request with
more care/custom handling, so you'd like to do a e.g. http_get request
separately from the rad block, do some error checking perhaps, and
*then* pass the json blob body to a display block for the json algo.

That's exactly what we allow with this commit. Previously, for a
'display' block, you could not specify a source. Now you can:

  a = { ... some dict data ... }
  display a:
    fields Name, ID

The src is optional -- if specified, Rad will do the json algo on the
given data to populate its fields before displaying them. If omitted, we
continue the current behavior of just assuming the given fields
*already* have data, and we'll try to display them.

This commit introduces a new implementation concept: RslTypeVisitor.
Right now, wherever we need to handle different types differently, we do
a switch on the RslValue's type and put the different handling into
different case types. This has a couple of related downsides:

1. It becomes hard to find all the places where were fork on the data
   type and have different handling for different types.
2. Furthermore, if we want to add more types in the future, it becomes a
   challenge to find all the places in code we need up to update.

By introducing a visitor, and if we consistently use it everywhere
instead of switches (that complete refactor is tbc), we can more easily
find these locations in code. The API of this visitor pattern may change
in future.