Add casting functions: 'int()' , 'float()'
There are some functions e.g. "rand_int()" which does not accept ints,
so even tho you have what is functionally an int e.g. 2.0, it might be
rejected because it's still a float underneath the hood.

An argument can be made that said functions should just realize that the
given arg is functionally an int, but I'm not convinced, yet. But
meanwhile, it seems reasonable to add a couple more casting functions to
join str().