Make round(n) and round(n, 0) return int
Right now, it returns a float. This is okay on one hand (nice to have a
consistent return type), but also, it can be unintuitive that e.g.
"round(2.2)" doesn't return back an integer.

Given RSL's role as a scripting language, let's be practical and give
the people what they want :) An int.