RLS: Give error diagnostics for unknown functions
A frequent problem, even for me, is remembering all the functions
correctly. I may mistype or misremember, but the RSL LSP server doesn't
know that the function I typed doesn't actually exist.

That changes with this commit. The flow is not too complex:

I've added a new 'main' function under function-metadata/extract.go
which is run as part of the RLS 'build' step. This main references Rad's
own internal "FunctionsByName" constant, reads out all the names, and
writes them to a file 'functions.txt'. The build step then moves this
file to an 'embedded' folder where it can be embedded in the language
server binary.

On startup, RLS then loads this embedded file, populates a set of valid
names, and when checking source code for errors/diagnostics, it will
consult this set of valid function names. Any unknown functions will be
marked as errors in the user's editor.

As part of this, we make a few changes in the RTS repo. This includes
moving the lang_consts.go file from this Rad repo into the RTS repo so
that we can reference the constants from there as well, so we need to
update usages all over Rad.