Remove function 'gen_nanoid'
After adding 'gen_stid', nanoid is actually kind of redundant, if my
understanding of it is correct.

With STID, you can disable the time component by setting the time
granularity to 0, and if you extend the 'num_random_chars' to 21 (same
as nanoid), you have effectively the same thing.

Let's remove nanoid - it's redundant.