Implement new concept: 'stash'
The previous commit implemented some basics around giving Rad a home
directory. This commit builds on that by allow scripts to store data
under this 'rad home'. Specifically, we introduce the idea of 'stashes'.

A stash is basically a namespace of data stored under rad's home as
files. Each stash has an ID. Stashes, as of this commit, have two
structures:

  ~/.rad/stash/<stash id>/state.json  << state file
  ~/.rad/stash/<stash id>/files/*     << directory for stash files

In order for a script to interact with a stash, it must invoke the
'set_stash_id()' function, which takes a stash ID as input.

If a script wants its own stash that others won't interfere with, it's
recommended that you give it a uniquely generated ID. For example:

  set_stash_id("1u509FxpsyS")

The Rad interpreter will store that this ID has been set, and will use
it for further stash-related operations. This includes the following
functions:

- load_state: load state.json from the stash into a map
- save_state: save the given state map into state.json

This will commonly look like:

  set_stash_id("1u509FxpsyS")
  state = load_state()
  defer save_state(state)
  ... read/modify state in remainder of script ...

- get_stash_dir: returns an absolute path to paths in the stash dir
- load_stash_file: reads the contents of a file under '<stash>/files/'
  and returns it. If it doesn't exist, creates the file, and returns the
  default arg given instead.
- write_stash_file: writes the given string content to the given
  relative path under '<stash>/files'.

All this functionality empowers RSL developers to more easily write
scripts with persistent state between invocations. While it was
previously possible by rolling your own persistent state using read_file
and write_file (and generally the fact that RSL is a turing complete
language), we want RSL to be first in class when it comes to making it
easy for developers to implement these sorts of functionalities, and let
them lean on RSL to have done the hard work, and exposing easy-to-use
tools for them.

There will be further work pushing this area.

Todos remaining after this commit:

- Unit tests
- Load stash should signal to user if it created the file or not
- Update docs / func ref
- 'Ensure' functions