Improve number precision preservation in json deserialization
For large numbers, approaching the limit of what can be stored in 64
bits, we may currently (prior to this commit) lose some precision when
they're deserialized from JSON.

For example, 38123123123123123 will become 38123123123123120 because the
former cannot be accurately represented in Go's float64 type, which is
what we deserialize into by default.

Realistically, we'll probably need to implement a sort of 'big decimal'
functionality, either explicitly for users to use, or behind-the-scenes
(tbd on how we'd do that in a performant way).

As a quick bandaid/improvement however, we'll deserialize numbers into
strings (json.Number) and then decide later if we should convert it into
an int64 or float64. We do this simply based on whether or not there's a
decimal in the number.

This should mean we can now correctly deserialize 64-bit integers that
cannot be represented as 64-bit floating numbers.