Add second return value 'existed' in load_state and load_stash_file
We probably needed *some* indicator of whether a state or stash file
already existed before the 'load' function created it. This commit adds
that as a second optional output for each.

We *could* make this another key in the map already returned by
load_stash_file, but this would be inconsistent with load_state, so we
keep them consistent this way by both returning a value.