Add functions: 'hash' & base16/base64 encode/decode
Notes on a few decisions:

- 'hash' as a single function with 'algo' as a named arg

The alternative here is to have dedicated functions for the different
algos e.g. md5(), sha1(), etc. I don't have a strong opinion on that, so
I decided to go with this low-API-surface option instead. We can easily
add more hashing algorithms with this shared 'hash' function.

The function defaults to 'sha1' rather than always requiring an explicit
algorith to be set. 'sha1' serves the usecase of being relatively fast
and the output not *that* long, so if someone just needs to quickly hash
some contents (without a care for security), hash() is a convenient way
to do that.

- encode/decode 'base16' instead of 'hex'

I don't do enough hex-stuff to have a *strong* preference here, and LLMs
were suggesting I stick with 'hex' naming, and so my choice here may be
something I go back on. For now though, I'm sticking with what's more
intuitive to me, and that's a consistent 'baseXX' naming scheme.

- Exclusion of a shorter-output hash function

Originally I had intended to include a 'crc32' option, but honestly it's
so trivial to do

  hash("my text")[:10]

if you just want a shorter hash. Taking this slice is a reasonable and
safe thing to do (understanding the security/collision implications), so
we keep it simple and will just expect users to do this instead.