Switch nwidger/jsoncolor to amterp/jsoncolor
I've forked the original because it appears unmaintained (unaddressed PR
for 2+ years). Said PR is also exactly what I'm after. My fork undoes a
broken commit and fixes what it aimed to fix correctly, and so by
switching to it, we now have correctly-formatted colored JSON output in
Rad, hurrah!

You can see the now-correct formatting in the tests that are updated as
part of this commit (as they will fail otherwise!).