Add func 'map'
Map takes a lambda, as we now support as of the previous commit. If used
on maps, the lambda is expected to take both a key and a value. For
lists, we don't unpack lists of lists. We may or may not allow that in
future, I've not thought much about it yet.

One rough edge in this implementation is that PosArg takes a node, and
we need to create these PosArgs for the sake of invoking the lambda, but
there is no node for those args (they're stored as values in the list
being map'd). We simply give it the lambda node itself, so that the
error pointer points at where things went wrong.