Implement function 'load'
Copy + paste of docs:

---

Loads a value into a map. Returns the mapped value.

 load(map, key, loader: fn() -> any, reload: bool?, override: any?) -> any

- Examples:
  - `load(m, k, loader)`
    - If `m` does not contain `k`, `loader` is run to calculate a value.
      This value is put into the map under `k` and returned.
    - If `m` contains `k`, `loader` is ignored, and the existing value
      is returned.
  - `load(m, k, loader, reload=true)`
    - Regardless of if `m` already contains `k`, `loader` is invoked and
      its value is put into the map for `k` and returned.
  - `load(m, k, loader, override=myvalue)`
    - Regardless of if `m` already contains `k`, if `myvalue` is a
      truthy value, then it is put into `m` under `k` and returned and
      `loader` is ignored.

`reload` cannot be true with `override` is truthy, that will return an
error.

---

This gets us a lot closer to completing the 'stash' feature in rad.
load() is envisioned as being the go-to way to do first-time loading
(you can invoke load() and pass a lambda using input() for first-time
setup) of configuration and script state. The 'override' and 'reload'
named args then offer options for redoing configuration beyond
first-time setups. This does put some onus on script writers to offer
reconfiguration options, so more may come so that script *users* have
some recourse for resetting a script's state and just starting over and
*forcing* the first-time setup experience for a script.

One wrinkle in the implementation is when both 'reload' and 'override'
are configured. This has convinced me that we need a null type in RSL,
as much as I've been trying to avoid it.