Define built-in functions as referenceable symbols
Some recent commits added fn / lambda support to RSL. This involved
defining them as symbols as part of the interpreter's environment. This
meant you could store a function as a variable and pass it to other
functions e.g.

  to_upper = fn(x) x.upper()
  mylist.map(to_upper)

However, you couldn't yet pass built-in functions around the same way,
i.e. this was not possible because 'upper' was not an actual symbol in
the environment, but *had* to be invoked wherever it was referenced

  mylist.map(upper)

This commit changes that. We join together more of the code path between
in-built and custom functions. Both are now defined as RslFn values in
the environment that can be referenced. When we create the interpreter,
as a initialize step, we go through all the in-built functions and add
them to the environment as RslFn values.

One detail we remember to update is the shell export functionality: we
simply skip over any RslFn values in the environment as they're not
possible/meaningful to export to a bash environment.

Behavior-wise, built in functions still have all their same validation
applied, etc.