Print short script usage for '-h' (excluding global flags)
If you simply invoke rad on a script with -h, we'll give the 'short'
version of its usage, which is just that script's specific usage -- none
of the global rad arguments.

Also, while here, don't print a 'Script args' header if the script
doesn't have args -- oversight that's survived this far.