Add null type
I've been trying to avoid this since the start, but there's no way. The
most compelling argument is probably that we're aiming to have type
parity with JSON, and since it has a null type, so should we.

But also practically speaking, when scripting, nulls *are* useful to
indicate an absence of something, at least as far as being a
lighter-weight version of optionals.

There are more commits to follow on this subject, this just does the
basics.