Null: Make undefined args null
If you have a script like this

  args:
    aaa string
    bbb string
    aaa mutually excludes bbb

and invoke this with '--aaa hi' i.e. define 'aaa' but not 'bbb', then
prior to this commit, 'bbb' simply is not part of the interpreter's
environment. The only way for the user to interact with 'bbb' to e.g.
check if 'aaa' or 'bbb' were given is to use the 'is_defined("bbb")'
function, but is not ideal.

Now that we have nulls, let's just make undefined args null. This makes
them a part of the environment and you can do e.g. truthy/falsy checks
to determine what was given or not.