Implement falsy coalescing
This is a similar idea to 'null coaliscing', but for truthy/falsy-ness.
Not sure if the term already exists but this is probably the best name
for it.

The idea is you can do this:

a = a or b or c

and 'a' will be equal to the first truthy value. It's a way to chain
potentially falsy values and get out a truthy one if it exists.

This also works for nulls since they're falsy. This might mean we don't
need a null coalescing operator '??' just yet, will wait and see if the
need arises.