Redo stash id metadata macro
I didn't like the previous take, and it also turned out to be difficult
for TextMate bundles to properly parse.

Let's do it differently. Instead of this

  ---
  Some docs.
  ---
  script_id = asdf
  ---

It's now

  ---
  Some docs.
  @script_id = asdf
  ---

i.e. any

 @<identifier> = <value>

Lines at the end of a file header will be interpreted as a metadata
macro and be excluded from the actual file header contents for
script usage.

We remove the 'set_stash_id' function, as this metadata macro is now the
new way to do it.

Also correct some erroneous 'Printf' renames from a previous commit.