Args: wire through optionality
We're adding the ability to specify an arg as optional with '?' in the
syntax. This was once in the language but didn't fit well, but it's
being added back in now that we have a null type, and it fits much
better.

A follow up commit will update & test handling for relational
constraints.