Fix table sizing & truncating when using ANSI coloring
The current logic wasn't properly ignoring ANSI escape codes when sizing
columns. Additionally, when truncating, we would sometimes truncate away
the escape codes themselves, meaning we'd have the opening for an escape
code and then get rid of the ending, making the code apply to undesired
text.

This commit fixes both these issues. Part of the fix lies in the go-tbl
library, so we also upgrade to the version released there.