Allow get_path to expand ~, add --state to 'stash' command
It can be useful to call get_path on a potential path that doesn't
actually exist, but then get the full path back still. Let's always put
that into the map. However, we previously used its emptiness to signify
that the path does not actually exist. So let's add a 'exist' field into
the map as well for this purpose.

Also add a --state flag to the stash command which simply prints the
state.json.