Implement initial 'rad check' command
 rad check <script>

Reports diagnostics, including errors, hints, lints, etc.

The meat of the logic behind this checker lives in rts, where it can
also be leveraged by the LSP. The idea is that all the issues that the
LSP can report will be reportable via this checker, and vice versa.

RslErrors have been moved to rts as well, so that the checker can refer
to them.

TODO:

- Improve formatting
- Implement checker Opts
- Implement --format=json
- Move 'unknown functions' check from LSP to checker
- Avoid double parse in Checker in LSP (see TODO comment)