Do not allow bool args to be provided positionally
Prior to this commit, it was possible to provide bool args positionally.
Tbh, this always felt a little weird. You could enter '1' or 'true', for
example. But I think we should just keep it simple and expect users to
pass the flag itself e.g. -d or --debug.

Specifically, we exclude bools from the positional arg logic, and update
the usage string generation to reflect this.

We also add [OPTIONS] onto the ending of usage summaries, as appears to
be standard practice. To make the 'global flags' more clearly included
in that, we switch that term to "global options" instead. This also
means updating the 'enable_global_flags' macro to
'enable_global_options'.