Implement 'catch' error syntax
Tricky change. We are basically implementing a custom, terse 'catch'
syntax into Rad. For example, if foo() can return an error:

 a = foo()

Here, we will throw an error if foo() returns an error i.e. 'a' will not
get defined. If the error is thrown at the top level, the script exits
on the spot with the associated message.

However, you can catch the error and save it as a variable:

 a = catch foo()

We are effectively returning a union type of the 'correct' return type
and 'error'.

The syntax features propagation between layers of functions. We achieve
the propagation via a panic-recover in our interpreter's Go
implementation. I spent a bit of time thinking about if we could come up
with alternatives, but I couldn't think of anything that wouldn't be
substantially more complicated.

We basically do the 'should we catch or throw' check whenever we've just
evaluated the result of a call.

A requirement of getting this change working is a change in function
semantics. Rather than Rad having true 'multiple return values' for
functions, we switch to a simpler model of (functions always return one
value, this this may be a tuple containing multiple values). This allows
us to return a tuple, or a single error, as a sort of union. Future
commits will complete this transition.

This commit also introduces the special 'RadError' type itself. We aim
for it to behave pretty much exactly like a string, in Rad semantics. We
include an 'error()' function to create them. There's a bit of special
handling for the error function, in that it is the only function in Rad
from which an error can be returned without us throwing it (the
alternative is just kinda silly i.e. 'a = catch error("what")' being
required.

We may want to use this 'catch' approach for shell command invocation as
well.