Revamp assignment: introduce destructuring
The rules we implement for destructuring are *very* lax. I think too
much so, we may tighten them. Demonstrated through switch statements
(which have the same rules as functions when returning values for
assignment):

 a = switch foo:
     case bar -> [1]     // a = [1]
     case quz -> [2, 3]  // a = [2, 3]

 [a, b] = switch foo:
     case bar -> [1]     // a = 1 | b = null
     case quz -> [2, 3]  // a = 2 | b = 3

Those two cases are pretty clear and mostly uncontroversial.
Next is where it gets trickier.

 a = switch foo:
     case bar -> 1     // a = 1
     case quz -> 2, 3  // a = [2, 3]

 [a, b] = switch foo:
     case bar -> 1     // a = 1 | b = null
     case quz -> 2, 3  // a = 2 | b = 3

My reasoning at this point is that this will only hold for cases without
typing. When we begin enforcing typing, then in the latter case, then
you'd write either `-> list` or `-> list[int]`, so the 'bar' case would
imply [1] is really what's being returned. OR we'd force users to return
[1] explicitly, I haven't decided. But in any case, it's clear what the
function is returning, and that we can destructure it. Without the
typing, things are lax, one return or branch may return 1 value, another
maybe a tuple of values. We are trying to be a productive/forgiving
language, though admittedly, this may be too much, as I said at the
start.

One complication is for a function like 'load_state'. With multiple
return values (MRV), we can conveniently return the state completely
separately from the 'existing' bool var to indicate if the state already
existed or not. However, now, we need to either return a tuple
containing the state and that bool, or add a 'existing' flag as a field
into the state map. But the latter feels bound to go wrong, as we're
interfering with user state. The issue with the former is just that it's
arguably a little verbose for a little-used case. It'd be good for us to
offer such a flag, but if doing so requires the user to always
destructure? I'm not sure. For now, we just don't return that flag, I'll
think more on it.

While here, we also prevent the 'print shell exports' functionality from
exporting PATH (quick win).