Write out Rad signatures for in-built functions
Been thinking about and developing what Rad's syntax for typing will
be. A good way to do that is to try it out on all our in-built
functions. Not only that, these declarations will likely double for all
our typing metadata and type checking in 'rad check'-related
functionality. It's also a good way to try and ensure that our syntax is
rich and expressive enough that I don't use unique Go implementation
that won't be available to users.

More to come:

- Update tree sitter syntax for 'fn types' (see e.g. load())
- Implement type checking
- Leverage these "stub" signatures to type check in-built functions