Update interpreter to function with new syntax
We've made various changes in the tree sitter grammar, this commit
updates the interpreter to work with them. There are several concrete
improvements, and some changes are setup for future functionality (type
checking).

1. Rename string type to str

It's shorter, is still clear, and saves us 3 chars each time it's
written, which will be particularly handle when writing function types.

2. Allow return/yields at any point in blocks

These were previous explicit statements expected at the end of switch
cases or functions, which meant you could't use them inside nested if
statements. Now you can.

This involves adding 'EvalResult' and 'CtrlKind' as concepts in the
interpreter.

Some bugs are expected, we will iron them out as we continue.