Implement runtime function type checking
The tree sitter grammar has been extended to allow for typing. Unlike
Python, we will enforce these types at runtime, and eventually have a
best-effort type checker. The runtime checking is quite a bit easier,
and allows us to put in place a lot of code we'll later re-use for the
static type checker.

The goal is for our function typing particularly to be very expressive,
and to allow fully typing our in-built functions as well. Making
built-in functions rely on our generate type checking will come in a
future commit.

I noticed we also allow duplicate named args, we now detect that and
stop it.