Add more function typing tests
- any
- optional
- unions
- varargs, almost

For varargs, I've realized that our syntax doesn't work. For example,
let's say we have a function

  fn foo(x: (*int)|int):
    pass

That's currently a completely valid typing. If we invoke 'foo(1)'
however, it becomes *functionally ambiguous*. If it were just *int as
the type, we'd wrap that 1 in a list i.e. pass [1]. But since it's
union'd with 'int', passing 1 as-is is also acceptable! Which should we
choose?

We could create a rule in the language for this scenario that the *int
takes precedence and we pass a list. But it's just a confusing situation
and language caveat. Let's switch to what other languages do, now that
we've learnt this lesson the hard way:

  fn foo(*x: int)

i.e. turn the vararg into a *parameter* modifier rather than a type
modifier. That's not in this commit, but will come.