Change variadic asterisk to be a parameter modifier, not type modifier
I came to the realization that the variadic syntax e.g. 'x: *int' was
not going to work, as there was room for ambiguity in calls like foo(1)
if the typing was e.g. 'fn foo(x: *int|int)' (should the x be 1 or
[1]?).

What other languages do, and I think at least in part for this reason,
is make the variadic asterisk* a modifier of the parameter name rather
than the type i.e. '*x: int'.

That's what we do as well, in this commit.

That means that the [*int] syntax no longer makes sense, so we change
that back to the ol' int[] syntax.

We also complete our typing tests.