Validate built-in functions using Rad syntax signatures
Previous to this commit, I'd rolled a few Go-implemented iterations of a
framework for writing built-in functions, and allowing them to concisely
specify their own signature, in a way. e.g. expected types, num args,
return values, etc.

Now that we have a rich typing system in Rad, and also because it'd be
good to force our built-in functions to behave like Rad functions, we
can replace these framework with Rad signatures.

See signatures.go.

We basically write a series of function signatures in this file, and
parse each of them to derive a TypingFnT, which is also what custom Rad
functions use to apply typing validation.

One big concern is that this is actually quite slow. In this commit, we
have 115 signatures we compile on startup every time. This is very
wasteful, adding a ~7 milli startup cost to every Rad script, which is
clearly something we'll want to address. Need to think of a better way.

There's also a little bit of trickiness around evaluating the TS nodes
in the signatures, as the nodes are tied to their original src code
strings, not the script that's being run, hence this 'WithTmpSrc'
solution. A little hacky, hopefully doesn't prove too annoying.

We temporarily remove the golines step from our build scripts, because
it makes the signatures.go file quite ugly to read. Need to find a way
to exclude this file from it.