Display just key in 'pick' selection
Previously, we were displaying

  <value> (<key>)

I don't know for sure why, I think to improve the matching/filtering.
It can read a little strangely though, and I don't think it works well
for if values are complex types.

Let's just display the keys. Might hit a reason to change it again.