Initial 'commands' implementation
This is a bit of a bare initial implementation. Missing testing,
better usage string, etc. The implementation is also probably not final,
it seems plausible we'll replace this syntax or implementation once
imports are implemented, but for now, I wanna get this feature
available as this is a common use case.

Not super well tested manually either, so I don't release this under a
version yet.