Args: Automatically replace underscores with hyphens in names
It's not uncommon to have multi-word arguments, which likely takes the
form of writing them with underscores e.g.

  args:
    first_name str

But then the user currently has to manually rename to use a hyphen
instead, since an underscore is highly unusual for users:

  args:
    first_name "first-name" str

Let's curate this a bit by making a reasonable assumption that we should
do this by default. If users don't want it, they can always use the
rename to bring the underscore back.