Allow sort() to sort parallel arrays
You often end up with parallel arrays in Rad, typically representing
columns of values.

It'd be useful to be able to sort them in parallel.

sort() now allows that. You can give it a variable number of lists to
sort, and if they're of the same length, it will sort the primary list
and reflect the ordering in the other lists it was given.