Introduce magefile.go
https://magefile.org/

See #28 for some discussion.
Magefiles are a kind of substitute for Makefiles that should be more
cross-platform-friendly as they can utilize Go to implement build steps.

This commit is mostly a port of the existing Makefile. A future commit
will convert these to using Go APIs instead of unix commands, for 'true'
cross-platform compatibility.