Merge rad/rts/lsp-server modules
Previous commit converted fully to workspaces, but I didn't like what
the resulting workflow would be. In order to maintain correct 'requires'
version references for rts in lsp-server for example, you would need to
first commit 'rts' changes, and then reference the resulting hash in the
lsp-server module and commit those changes separately.

We'll keep the previous commit just for posterity, but this commit here
quickly follows and pivots us to just try having a single module. I'll
be testing how this affects user installation in practice, we may or may
not revert if there are other issues I'm not seeing right now.