http functions: add 'json' named arg
The current 'body' arg kinda behaves like 'json' should i.e.
json-encodes whatever you pass to it. This means we currently can't POST
e.g. plain text as-is with Rad, since we wrap it in quotes, to make it
json-compatible.

Let's separate out a 'json' named arg which does this. If no headers
specified, we also automatically put the Content-Type correctly.

'body' will now just treat whatever you give it as-is.