Replace pflag with Ra
spf13/pflag is great, but Rad has been straining it for quite some time.
It's too opinionated about POSIX. For Rad, we'd like to allow users to
breach POSIX compliance at their discretion, such as having varargs to
their scripts. pflag does not support this.

Hence the implementation of Ra (https://github.com/amterp/ra). This
commit migrates Rad entirely from pflag to Ra. This is quite a big lift,
and changes some nuanced behaviors of Rad. I will no doubt uncover
many-a-bug from this point onwards, so will be an ongoing thing.

Ra is built on principles we've already been using in Rad, for example
typed args, dual-nature of args (positional/flag), etc, but the idea is
that it can support more, which we will later leverage in Rad e.g. var
args.