fix: convert arg constraint targets to external name format
Rad was misusing the Ra lib. Rad transforms the names of args slightly
(right now just underscores to hyphens) and gives these to Ra, but it
was not doing it for constraints e.g. in

  aa_bb requires cc_dd

Ra would be told 'aa-bb' requrires 'cc_dd', but the latter would've been
registered as 'cc-dd', and we get a 'undefined flag' error from Ra
later.

This commit just ensures we also run 'cc_dd' through the 'externalizer'.